moved VC code out of main src
