+++ demo/6_resume_credential.js
@@ -1,6 +1,6 @@
 const { PrivateKey, Client, FileId } = require("@hashgraph/sdk");
-const { HcsVc } = require("../dist");
-const { OPERATOR_ID, OPERATOR_KEY, REVOCATION_LIST_FILE_ID, ISSUER_DID, ISSUER_PK, TOPIC_ID } = require("./.env.json");
+const { HcsRl } = require("../dist");
+const { OPERATOR_ID, OPERATOR_KEY, REVOCATION_LIST_FILE_ID } = require("./.env.json");
 
 async function main() {
     /**
@@ -10,9 +10,6 @@ async function main() {
     client.setOperator(OPERATOR_ID, OPERATOR_KEY);
 
     const hcsVc = new HcsVc(
-        ISSUER_DID,
-        PrivateKey.fromString(ISSUER_PK), // this is to sign message
-        TOPIC_ID,
         PrivateKey.fromString(OPERATOR_KEY), // this is to sign transaction
         client
     );
