+++ script/vc/index.ts
@@ -1,0 +1,5 @@
+import { W3CCredential } from "./w3c-credential";
+import { VCError, VCErrorCode } from "./vc-error";
+import { VC } from "./vc";
+
+export { VC, VCError, VCErrorCode, W3CCredential };
