+++ script/vc/utils/jwt.ts
