+++ src/index.ts
@@ -1,6 +1,4 @@
-import { HcsVc } from "./identity/hcs/vc/hcs-vc";
-import { W3CCredential } from "./identity/hcs/vc/w3c-credential";
-import { VCError, VCErrorCode } from "./identity/vc-error";
+import { HcsRl } from "./identity/hcs/rl/hcs-rl";
 import { VcStatus } from "./identity/vc-status";
 
-export { HcsVc, VCError, VCErrorCode, VcStatus, W3CCredential };
+export { HcsRl, VcStatus };
